PYTHON/requirements.txt
numpy 
pillow
matplotlib
pandas